+++ ChangeLog.md
@@ -3,6 +3,10 @@ XP Framework Core ChangeLog
 
 ## ?.?.? / ????-??-??
 
+### Features
+
+* Merged PR #369: Replace spl_object_hash() with spl_object_id() - @thekid
+
 ## 12.11.0 / 2026-06-21
 
 ### Features